+++ samples/idtokens-serverless.js
@@ -22,6 +22,12 @@ function main(
   url = 'https://service-1234-uc.a.run.app',
   targetAudience = null
 ) {
+  if (!targetAudience) {
+    // Use the target service's hostname as the target audience for requests.
+    // (For example: https://my-cloud-run-service.run.app)
+    const {URL} = require('url');
+    targetAudience = new URL(url).origin;
+  }
   // [START google_auth_idtoken_serverless]
   // [START cloudrun_service_to_service_auth]
   // [START run_service_to_service_auth]
@@ -29,17 +35,15 @@ function main(
   /**
    * TODO(developer): Uncomment these variables before running the sample.
    */
-  // const url = 'https://TARGET_URL';
-  // let targetAudience = null;
+  // Example: https://my-cloud-run-service.run.app/books/delete/12345
+  // const url = 'https://TARGET_HOSTNAME/TARGET_URL';
+
+  // Example: https://my-cloud-run-service.run.app/
+  // const targetAudience = 'https://TARGET_HOSTNAME/';
   const {GoogleAuth} = require('google-auth-library');
   const auth = new GoogleAuth();
 
   async function request() {
-    if (!targetAudience) {
-      // Use the request URL hostname as the target audience for requests.
-      const {URL} = require('url');
-      targetAudience = new URL(url);
-    }
     console.info(`request ${url} with target audience ${targetAudience}`);
     const client = await auth.getIdTokenClient(targetAudience);
     const res = await client.request({url});