+++ test/test.baseexternalclient.ts
@@ -82,6 +82,24 @@ describe('BaseExternalAccountClient', () => {
     client_id: 'CLIENT_ID',
     client_secret: 'SECRET',
   };
+  const externalAccountOptionsWorkforceUserProject = Object.assign(
+    {},
+    externalAccountOptions,
+    {
+      workforce_pool_user_project: 'workforce_pool_user_project',
+      audience:
+        '//iam.googleapis.com/locations/global/workforcePools/pool/providers/provider',
+      subject_token_type: 'urn:ietf:params:oauth:token-type:id_token',
+    }
+  );
+  const externalAccountOptionsWithClientAuthAndWorkforceUserProject =
+    Object.assign(
+      {
+        client_id: 'CLIENT_ID',
+        client_secret: 'SECRET',
+      },
+      externalAccountOptionsWorkforceUserProject
+    );
   const basicAuthCreds =
     `${externalAccountOptionsWithCreds.client_id}:` +
     `${externalAccountOptionsWithCreds.client_secret}`;
@@ -103,6 +121,12 @@ describe('BaseExternalAccountClient', () => {
       service_account_impersonation_url: getServiceAccountImpersonationUrl(),
     },
     externalAccountOptionsWithCreds
+  );
+  const externalAccountOptionsWithWorkforceUserProjectAndSA = Object.assign(
+    {
+      service_account_impersonation_url: getServiceAccountImpersonationUrl(),
+    },
+    externalAccountOptionsWorkforceUserProject
   );
   const indeterminableProjectIdAudiences = [
     // Legacy K8s audience format.
@@ -245,6 +269,64 @@ describe('BaseExternalAccountClient', () => {
       }
     });
 
+    const invalidWorkforceAudiences = [
+      '//iam.googleapis.com/locations/global/workloadIdentityPools/pool/providers/provider',
+      '//iam.googleapis.com/locations/global/workforcepools/pool/providers/provider',
+      '//iam.googleapis.com/locations/global/workforcePools//providers/provider',
+      '//iam.googleapis.com/locations/global/workforcePools/providers/provider',
+      '//iam.googleapis.com/locations/global/workloadIdentityPools/workforcePools/pool/providers/provider',
+      '//iam.googleapis.com//locations/global/workforcePools/pool/providers/provider',
+      '//iam.googleapis.com/project/123/locations/global/workforcePools/pool/providers/provider',
+      '//iam.googleapis.com/locations/global/workforcePools/pool/providers',
+      '//iam.googleapis.com/locations/global/workforcePools/workloadIdentityPools/pool/providers/provider',
+      '//iam.googleapis.com/locations/global/workforcePools/pool/providers/',
+      '//iam.googleapis.com/locations//workforcePools/pool/providers/provider',
+      '//iam.googleapis.com/locations/workforcePools/pool/providers/provider',
+    ];
+    const invalidExternalAccountOptionsWorkforceUserProject = Object.assign(
+      {},
+      externalAccountOptionsWorkforceUserProject
+    );
+    const expectedWorkforcePoolUserProjectError = new Error(
+      'workforcePoolUserProject should not be set for non-workforce pool ' +
+        'credentials.'
+    );
+
+    invalidWorkforceAudiences.forEach(invalidWorkforceAudience => {
+      it(`should throw given audience ${invalidWorkforceAudience} with user project defined in options`, () => {
+        invalidExternalAccountOptionsWorkforceUserProject.audience =
+          invalidWorkforceAudience;
+
+        assert.throws(() => {
+          return new TestExternalAccountClient(
+            invalidExternalAccountOptionsWorkforceUserProject
+          );
+        }, expectedWorkforcePoolUserProjectError);
+      });
+    });
+
+    it('should not throw on valid workforce audience configs', () => {
+      const validWorkforceAudiences = [
+        '//iam.googleapis.com/locations/global/workforcePools/workforcePools/providers/provider',
+        '//iam.googleapis.com/locations/global/workforcePools/pool/providers/provider',
+        '//iam.googleapis.com/locations/global/workforcePools/workloadPools/providers/oidc',
+      ];
+      const validExternalAccountOptionsWorkforceUserProject = Object.assign(
+        {},
+        externalAccountOptionsWorkforceUserProject
+      );
+      for (const validWorkforceAudience of validWorkforceAudiences) {
+        validExternalAccountOptionsWorkforceUserProject.audience =
+          validWorkforceAudience;
+
+        assert.doesNotThrow(() => {
+          return new TestExternalAccountClient(
+            validExternalAccountOptionsWorkforceUserProject
+          );
+        });
+      }
+    });
+
     it('should not throw on valid options', () => {
       assert.doesNotThrow(() => {
         return new TestExternalAccountClient(externalAccountOptions);
@@ -280,6 +362,16 @@ describe('BaseExternalAccountClient', () => {
   });
 
   describe('projectNumber', () => {
+    it('should return null for workforce pools with workforce_pool_user_project', () => {
+      const options = Object.assign(
+        {},
+        externalAccountOptionsWorkforceUserProject
+      );
+      const client = new TestExternalAccountClient(options);
+
+      assert(client.projectNumber === null);
+    });
+
     it('should be set if determinable', () => {
       const projectNumber = 'my-proj-number';
       const options = Object.assign({}, externalAccountOptions);
@@ -342,6 +434,65 @@ describe('BaseExternalAccountClient', () => {
   });
 
   describe('getProjectId()', () => {
+    it('should resolve for workforce pools when workforce_pool_user_project is provided', async () => {
+      const options = Object.assign(
+        {},
+        externalAccountOptionsWorkforceUserProject
+      );
+      const projectNumber = options.workforce_pool_user_project;
+      const projectId = 'my-proj-id';
+      const response = {
+        projectNumber,
+        projectId,
+        lifecycleState: 'ACTIVE',
+        name: 'project-name',
+        createTime: '2018-11-06T04:42:54.109Z',
+        parent: {
+          type: 'folder',
+          id: '12345678901',
+        },
+      };
+      const scopes = [
+        mockStsTokenExchange([
+          {
+            statusCode: 200,
+            response: stsSuccessfulResponse,
+            request: {
+              grant_type: 'urn:ietf:params:oauth:grant-type:token-exchange',
+              audience:
+                '//iam.googleapis.com/locations/global/workforcePools/pool/providers/provider',
+              scope: 'https://www.googleapis.com/auth/cloud-platform',
+              requested_token_type:
+                'urn:ietf:params:oauth:token-type:access_token',
+              subject_token: 'subject_token_0',
+              subject_token_type: 'urn:ietf:params:oauth:token-type:id_token',
+              options: JSON.stringify({
+                userProject: options.workforce_pool_user_project,
+              }),
+            },
+          },
+        ]),
+        mockCloudResourceManager(
+          options.workforce_pool_user_project,
+          stsSuccessfulResponse.access_token,
+          200,
+          response
+        ),
+      ];
+
+      const client = new TestExternalAccountClient(options);
+      const actualProjectId = await client.getProjectId();
+
+      assert.strictEqual(actualProjectId, projectId);
+      assert.strictEqual(client.projectId, projectId);
+
+      // Next call should return cached result.
+      const cachedProjectId = await client.getProjectId();
+
+      assert.strictEqual(cachedProjectId, projectId);
+      scopes.forEach(scope => scope.done());
+    });
+
     it('should resolve with projectId when determinable', async () => {
       const projectNumber = 'my-proj-number';
       const projectId = 'my-proj-id';
@@ -473,6 +624,122 @@ describe('BaseExternalAccountClient', () => {
         ]);
 
         const client = new TestExternalAccountClient(externalAccountOptions);
+        const actualResponse = await client.getAccessToken();
+
+        // Confirm raw GaxiosResponse appended to response.
+        assertGaxiosResponsePresent(actualResponse);
+        delete actualResponse.res;
+        assert.deepStrictEqual(actualResponse, {
+          token: stsSuccessfulResponse.access_token,
+        });
+        scope.done();
+      });
+
+      it('should use client auth over passing the workforce user project when both are provided', async () => {
+        const scope = mockStsTokenExchange([
+          {
+            statusCode: 200,
+            response: stsSuccessfulResponse,
+            request: {
+              grant_type: 'urn:ietf:params:oauth:grant-type:token-exchange',
+              audience:
+                '//iam.googleapis.com/locations/global/workforcePools/pool/providers/provider',
+              scope: 'https://www.googleapis.com/auth/cloud-platform',
+              requested_token_type:
+                'urn:ietf:params:oauth:token-type:access_token',
+              subject_token: 'subject_token_0',
+              subject_token_type: 'urn:ietf:params:oauth:token-type:id_token',
+            },
+            additionalHeaders: {
+              Authorization: `Basic ${crypto.encodeBase64StringUtf8(
+                basicAuthCreds
+              )}`,
+            },
+          },
+        ]);
+
+        const client = new TestExternalAccountClient(
+          externalAccountOptionsWithClientAuthAndWorkforceUserProject
+        );
+        const actualResponse = await client.getAccessToken();
+
+        // Confirm raw GaxiosResponse appended to response.
+        assertGaxiosResponsePresent(actualResponse);
+        delete actualResponse.res;
+        assert.deepStrictEqual(actualResponse, {
+          token: stsSuccessfulResponse.access_token,
+        });
+        scope.done();
+      });
+
+      it('should pass the workforce user project on workforce configs when client auth is not provided ', async () => {
+        const scope = mockStsTokenExchange([
+          {
+            statusCode: 200,
+            response: stsSuccessfulResponse,
+            request: {
+              grant_type: 'urn:ietf:params:oauth:grant-type:token-exchange',
+              audience:
+                '//iam.googleapis.com/locations/global/workforcePools/pool/providers/provider',
+              scope: 'https://www.googleapis.com/auth/cloud-platform',
+              requested_token_type:
+                'urn:ietf:params:oauth:token-type:access_token',
+              subject_token: 'subject_token_0',
+              subject_token_type: 'urn:ietf:params:oauth:token-type:id_token',
+              options: JSON.stringify({
+                userProject:
+                  externalAccountOptionsWorkforceUserProject.workforce_pool_user_project,
+              }),
+            },
+          },
+        ]);
+
+        const client = new TestExternalAccountClient(
+          externalAccountOptionsWorkforceUserProject
+        );
+        const actualResponse = await client.getAccessToken();
+
+        // Confirm raw GaxiosResponse appended to response.
+        assertGaxiosResponsePresent(actualResponse);
+        delete actualResponse.res;
+        assert.deepStrictEqual(actualResponse, {
+          token: stsSuccessfulResponse.access_token,
+        });
+        scope.done();
+      });
+
+      it('should not throw if client auth is provided but workforce user project is not', async () => {
+        const scope = mockStsTokenExchange([
+          {
+            statusCode: 200,
+            response: stsSuccessfulResponse,
+            request: {
+              grant_type: 'urn:ietf:params:oauth:grant-type:token-exchange',
+              audience:
+                '//iam.googleapis.com/locations/global/workforcePools/pool/providers/provider',
+              scope: 'https://www.googleapis.com/auth/cloud-platform',
+              requested_token_type:
+                'urn:ietf:params:oauth:token-type:access_token',
+              subject_token: 'subject_token_0',
+              subject_token_type: 'urn:ietf:params:oauth:token-type:id_token',
+            },
+            additionalHeaders: {
+              Authorization: `Basic ${crypto.encodeBase64StringUtf8(
+                basicAuthCreds
+              )}`,
+            },
+          },
+        ]);
+        const externalAccountOptionsWithClientAuth: BaseExternalAccountClientOptions =
+          Object.assign(
+            {},
+            externalAccountOptionsWithClientAuthAndWorkforceUserProject
+          );
+        delete externalAccountOptionsWithClientAuth.workforce_pool_user_project;
+
+        const client = new TestExternalAccountClient(
+          externalAccountOptionsWithClientAuth
+        );
         const actualResponse = await client.getAccessToken();
 
         // Confirm raw GaxiosResponse appended to response.
@@ -1296,6 +1563,55 @@ describe('BaseExternalAccountClient', () => {
 
         const client = new TestExternalAccountClient(
           externalAccountOptionsWithCredsAndSA
+        );
+        const actualResponse = await client.getAccessToken();
+
+        // Confirm raw GaxiosResponse appended to response.
+        assertGaxiosResponsePresent(actualResponse);
+        delete actualResponse.res;
+        assert.deepStrictEqual(actualResponse, {
+          token: saSuccessResponse.accessToken,
+        });
+        scopes.forEach(scope => scope.done());
+      });
+
+      it('should still pass workforce user project when no client auth is used', async () => {
+        const scopes: nock.Scope[] = [];
+        scopes.push(
+          mockStsTokenExchange([
+            {
+              statusCode: 200,
+              response: stsSuccessfulResponse,
+              request: {
+                grant_type: 'urn:ietf:params:oauth:grant-type:token-exchange',
+                audience:
+                  '//iam.googleapis.com/locations/global/workforcePools/pool/providers/provider',
+                scope: 'https://www.googleapis.com/auth/cloud-platform',
+                requested_token_type:
+                  'urn:ietf:params:oauth:token-type:access_token',
+                subject_token: 'subject_token_0',
+                subject_token_type: 'urn:ietf:params:oauth:token-type:id_token',
+                options: JSON.stringify({
+                  userProject:
+                    externalAccountOptionsWithWorkforceUserProjectAndSA.workforce_pool_user_project,
+                }),
+              },
+            },
+          ])
+        );
+        scopes.push(
+          mockGenerateAccessToken([
+            {
+              statusCode: 200,
+              response: saSuccessResponse,
+              token: stsSuccessfulResponse.access_token,
+              scopes: ['https://www.googleapis.com/auth/cloud-platform'],
+            },
+          ])
+        );
+
+        const client = new TestExternalAccountClient(
+          externalAccountOptionsWithWorkforceUserProjectAndSA
         );
         const actualResponse = await client.getAccessToken();