+++ test/test.oauth2.ts
@@ -893,7 +893,7 @@ describe('oauth2', () => {
       client.request({}, (err, result) => {
         assert.strictEqual(
           err!.message,
-          'No access, refresh token or API key is set.'
+          'No access, refresh token, API key or refresh handler callback is set.'
         );
         assert.strictEqual(result, undefined);
         done();
@@ -1106,19 +1106,30 @@ describe('oauth2', () => {
 
     [401, 403].forEach(code => {
       it(`should refresh token if the server returns ${code}`, done => {
-        const scope = nock('http://example.com')
-          .get('/access')
-          .reply(code, {
-            error: {code, message: 'Invalid Credentials'},
-          });
-        const scopes = mockExample();
+        const scopes = [
+          nock('http://example.com')
+            .get('/access')
+            .reply(code, {
+              error: {code, message: 'Invalid Credentials'},
+            })
+            .get('/access', undefined, {
+              reqheaders: {Authorization: 'Bearer abc123'},
+            })
+            .reply(200),
+          nock(baseUrl)
+            .post('/token', undefined, {
+              reqheaders: {'content-type': 'application/x-www-form-urlencoded'},
+            })
+            .reply(200, {access_token: 'abc123', expires_in: 1000}),
+        ];
         client.credentials = {
           access_token: 'initial-access-token',
           refresh_token: 'refresh-token-placeholder',
         };
-        client.request({url: 'http://example.com/access'}, () => {
-          scope.done();
-          scopes[0].done();
+
+        client.request({url: 'http://example.com/access'}, err => {
+          assert.strictEqual(err, null);
+          scopes.forEach(scope => scope.done());
           assert.strictEqual('abc123', client.credentials.access_token);
           done();
         });
@@ -1131,22 +1142,107 @@ describe('oauth2', () => {
           redirectUri: REDIRECT_URI,
           forceRefreshOnFailure: true,
         });
-        const scope = nock('http://example.com')
-          .get('/access')
-          .reply(code, {
-            error: {code, message: 'Invalid Credentials'},
-          });
-        const scopes = mockExample();
+        const scopes = [
+          nock(baseUrl)
+            .post('/token', undefined, {
+              reqheaders: {'content-type': 'application/x-www-form-urlencoded'},
+            })
+            .reply(200, {access_token: 'abc123', expires_in: 1000}),
+          nock('http://example.com')
+            .get('/access')
+            .reply(code, {
+              error: {code, message: 'Invalid Credentials'},
+            })
+            .get('/access', undefined, {
+              reqheaders: {Authorization: 'Bearer abc123'},
+            })
+            .reply(200),
+        ];
         client.credentials = {
           access_token: 'initial-access-token',
           refresh_token: 'refresh-token-placeholder',
           expiry_date: new Date().getTime() + 500000,
         };
-        client.request({url: 'http://example.com/access'}, () => {
-          scope.done();
-          scopes[0].done();
+
+        client.request({url: 'http://example.com/access'}, err => {
+          assert.strictEqual(err, null);
+          scopes.forEach(scope => scope.done());
           assert.strictEqual('abc123', client.credentials.access_token);
           done();
+        });
+      });
+
+      it('should call refreshHandler in request() on token expiration and no refresh token available', async () => {
+        const authHeaders = {
+          Authorization: 'Bearer access_token',
+        };
+        const scope = nock('http://example.com')
+          .get('/access')
+          .reply(code, {
+            error: {code, message: 'Invalid Credentials'},
+          })
+          .get('/access', undefined, {
+            reqheaders: authHeaders,
+          })
+          .reply(200, {foo: 'bar'});
+        const expectedRefreshedAccessToken = {
+          access_token: 'access_token',
+          expiry_date: new Date().getTime() + 3600 * 1000,
+        };
+        client.refreshHandler = async () => {
+          return expectedRefreshedAccessToken;
+        };
+        client.setCredentials({
+          access_token: 'initial-access-token',
+          expiry_date: new Date().getTime() - 1000,
+        });
+
+        client.request({url: 'http://example.com/access'}, err => {
+          assert.strictEqual(err, null);
+          scope.done();
+          assert.strictEqual(
+            client.credentials.access_token,
+            expectedRefreshedAccessToken.access_token
+          );
+          assert.strictEqual(
+            client.credentials.expiry_date,
+            expectedRefreshedAccessToken.expiry_date
+          );
+        });
+      });
+
+      it('should call refreshHandler in request() if no credentials available', async () => {
+        const authHeaders = {
+          Authorization: 'Bearer access_token',
+        };
+        const scope = nock('http://example.com')
+          .get('/access')
+          .reply(code, {
+            error: {code, message: 'Invalid Credentials'},
+          })
+          .get('/access', undefined, {
+            reqheaders: authHeaders,
+          })
+          .reply(200, {foo: 'bar'});
+        const expectedRefreshedAccessToken = {
+          access_token: 'access_token',
+          expiry_date: new Date().getTime() + 3600 * 1000,
+        };
+        client.refreshHandler = async () => {
+          return expectedRefreshedAccessToken;
+        };
+
+        client.request({url: 'http://example.com/access'}, err => {
+          assert.strictEqual(err, null);
+          scope.done();
+          assert.strictEqual(
+            client.credentials.access_token,
+            expectedRefreshedAccessToken.access_token
+          );
+          assert.strictEqual(
+            client.credentials.expiry_date,
+            expectedRefreshedAccessToken.expiry_date
+          );
         });
       });
     });
@@ -1338,14 +1434,132 @@ describe('oauth2', () => {
       assert.deepStrictEqual(info.scopes, tokenInfo.scope.split(' '));
     });
 
-    it('should throw if tries to refresh but no refresh token is available', async () => {
+    it('should call refreshHandler in getRequestHeaders() when no credentials but refreshHandler is available', async () => {
+      const expectedRefreshedAccessToken = {
+        access_token: 'access_token',
+        expiry_date: new Date().getTime() + 3600 * 1000,
+      };
+      client.refreshHandler = async () => {
+        return expectedRefreshedAccessToken;
+      };
+      const expectedMetadata = {
+        Authorization: 'Bearer access_token',
+      };
+      assert.deepStrictEqual(client.credentials, {});
+
+      const requestMetaData = await client.getRequestHeaders(
+        'http://example.com'
+      );
+
+      assert.deepStrictEqual(requestMetaData, expectedMetadata);
+    });
+
+    it('should call refreshHandler in getRequestHeaders() on token expiration and refreshHandler is available', async () => {
+      const expectedRefreshedAccessToken = {
+        access_token: 'access_token',
+        expiry_date: new Date().getTime() + 3600 * 1000,
+      };
+      client.refreshHandler = async () => {
+        return expectedRefreshedAccessToken;
+      };
       client.setCredentials({
         access_token: 'initial-access-token',
         expiry_date: new Date().getTime() - 1000,
       });
+      const expectedMetadata = {
+        Authorization: 'Bearer access_token',
+      };
+
+      const requestMetaData = await client.getRequestHeaders(
+        'http://example.com'
+      );
+
+      assert.deepStrictEqual(requestMetaData, expectedMetadata);
+    });
+
+    it('should return cached authorization header on getRequestHeaders() if not expired', async () => {
+      client.credentials = {
+        access_token: 'initial-access-token',
+        expiry_date: new Date().getTime() + 3600 * 1000,
+      };
+      const expectedMetadata = {
+        Authorization: 'Bearer initial-access-token',
+      };
+
+      const requestMetaData = await client.getRequestHeaders(
+        'http://example.com'
+      );
+
+      assert.deepStrictEqual(requestMetaData, expectedMetadata);
+    });
+
+    it('should throw on getRequestHeaders() when neither refreshHandler nor refresh token is available', async () => {
+      client.setCredentials({
+        access_token: 'initial-access-token',
+        expiry_date: new Date().getTime() - 1000,
+      });
+
       await assert.rejects(
         client.getRequestHeaders('http://example.com'),
         /No refresh token is set./
+      );
+    });
+
+    it('should call refreshHandler in getAccessToken() when neither credentials nor refresh token is available', async () => {
+      const expectedRefreshedAccessToken = {
+        access_token: 'access_token',
+        expiry_date: new Date().getTime() + 3600 * 1000,
+      };
+      client.refreshHandler = async () => {
+        return expectedRefreshedAccessToken;
+      };
+      assert.deepStrictEqual(client.credentials, {});
+
+      const refreshedAccessToken = await client.getAccessToken();
+
+      assert.strictEqual(
+        refreshedAccessToken.token,
+        expectedRefreshedAccessToken.access_token
+      );
+    });
+
+    it('should call refreshHandler in getAccessToken() on expiration and no refresh token available', async () => {
+      const expectedRefreshedAccessToken = {
+        access_token: 'access_token',
+        expiry_date: new Date().getTime() + 3600 * 1000,
+      };
+      client.refreshHandler = async () => {
+        return expectedRefreshedAccessToken;
+      };
+      client.setCredentials({
+        access_token: 'initial-access-token',
+        expiry_date: new Date().getTime() - 1000,
+      });
+
+      const refreshedAccessToken = await client.getAccessToken();
+
+      assert.strictEqual(
+        refreshedAccessToken.token,
+        expectedRefreshedAccessToken.access_token
+      );
+    });
+
+    it('should throw error if refreshHandler callback response is missing an access token', async () => {
+      const expectedRefreshedAccessToken = {
+        access_token: '',
+        expiry_date: new Date().getTime() + 3600 * 1000,
+      };
+      client.refreshHandler = async () => {
+        return expectedRefreshedAccessToken;
+      };
+      client.setCredentials({
+        access_token: 'initial-access-token',
+        expiry_date: new Date().getTime() - 1000,
+      });
+
+      await assert.rejects(
+        client.getAccessToken(),
+        /No access token is returned by the refreshHandler callback./
       );
     });
   });